Add a Directory.Exist check to DirectoryEx.SafeDelete to avoid unnecessary exceptions and retries
Remove the logic from TestGroupedStudiesTutorialDraft that had it running mzML data instead of WIFF data every other day. It seems stable with either at this point.

Cut back the extra retry pause for TryTwice failures in code coverage tests from 60sec to 5sec - it didn't seem to help

+++ pwiz_tools/Skyline/Util/UtilIO.cs
@@ -954,7 +954,10 @@ namespace pwiz.Skyline.Util
         {
             try
             {
-                Helpers.TryTwice(() => Directory.Delete(path, true), $@"Directory.Delete({path})");
+                if (path != null && Directory.Exists(path)) // Don't waste time trying to delete something that's already deleted
+                {
+                    Helpers.TryTwice(() => Directory.Delete(path, true), $@"Directory.Delete({path})");
+                }
             }
 // ReSharper disable EmptyGeneralCatchClause
             catch (Exception) { }
